# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "牧民健康应用API",
    "description": "牧民健康应用的后端API接口",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API服务器"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "timestamp": {
            "type": "integer",
            "description": "时间戳（毫秒）"
          }
        }
      },
...
      "Article": {
        "type": "object",
        "properties": {
          "articleId": {
            "type": "string",
            "description": "文章ID"
          },
          "title": {
            "type": "string",
            "description": "文章标题"
          },
          "title_mn": {
            "type": "string",
            "description": "蒙文标题"
          },
          "summary": {
            "type": "string",
            "description": "文章摘要"
          },
          "coverImage": {
            "type": "string",
            "description": "封面图片URL"
          },
          "author": {
            "type": "string",
            "description": "作者"
          },
          "viewCount": {
            "type": "integer",
            "description": "阅读量"
          },
          "category": {
            "type": "string",
            "description": "分类名称"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "标签列表"
          },
          "publishTime": {
            "type": "string",
            "format": "date-time",
            "description": "发布时间"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/articles": {
      "get": {
        "tags": ["健康文章"],
        "summary": "获取文章列表",
        "description": "获取健康文章列表",
        "security": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "分类名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "标签名称",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer",
                              "description": "总记录数"
                            },
                            "pages": {
                              "type": "integer",
                              "description": "总页数"
                            },
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Article"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/articles/{article_id}": {
      "get": {
        "tags": ["健康文章"],
        "summary": "获取文章详情",
        "description": "获取指定ID的文章详情",
        "security": [],
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/Article"
                            },
                            {
                              "type": "object",
                              "properties": {
                                "content": {
                                  "type": "string",
                                  "description": "文章内容（HTML格式）"
                                },
                                "content_mn": {
                                  "type": "string",
                                  "description": "蒙文文章内容（HTML格式）"
                                },
                                "relatedArticles": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "properties": {
                                      "articleId": {
                                        "type": "string",
                                        "description": "文章ID"
                                      },
                                      "title": {
                                        "type": "string",
                                        "description": "文章标题"
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "文章不存在"
          }
        }
      }
    },
...
  }
} 